pyproject.toml
[tool.poetry]
name = "cigarette-detection"
version = "0.1.0"
description = ""
authors = ["Tanya"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
Flask = "^2.0.2"
numpy = "^1.21.4"
torch = "^1.10.1"
opencv-python = "^4.5.4"
pandas = "^1.3.5"
Pillow = "^8.4.0"
PyYAML = "^6.0"
tqdm = "^4.62.3"
torchvision = "^0.11.2"
matplotlib = "^3.5.1"
seaborn = "^0.11.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
